+++ tests/Util/BufferTest.php
@@ -75,7 +75,7 @@ class BufferTest extends \PHPUnit_Framework_TestCase
     public function testSerialize()
     {
         $hex = '41414141';
-        $dec = (int)Math::hexDec($hex);
+        $dec = Math::hexDec($hex);
         $bin = pack("H*", $hex);
         $this->buffer = Buffer::hex($hex);
